sensor library update (MPU 6050 + ADXL 335)

+++ platform/sky/dev/adc-arch.h
@@ -9,7 +9,7 @@
 //DO NOT USE THESE FUNCTIONS IN APPLICATION LAYER
 void adc_arch_on(void);
 void adc_arch_off(void);
-unsigned short adc_arch_sample(void);
+unsigned int adc_arch_sample(void);
 int adc_arch_configure(unsigned char chan_num);
 
 #endif /* ADC_ARCH_H */
